08_personal_chat_assistant/requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.33.0
chromadb==0.4.24
pydantic==2.6.0
python-dotenv==1.0.1
